int -> long

+++ src/me/konstantinberkow/SchenksInd.java
@@ -15,14 +15,14 @@ public class SchenksInd {
      * @param p - нечетное прсотое
      * @return a - искомая хуйня
      */
-    public static int ind(final int x, final int g, final int p) {
+    public static long ind(final long x, final long g, final long p) {
         final int m = (int) (Math.sqrt(p) + 1); //"большие" шаги
         System.out.println("Big steps count: " + m);
 
-        final int step = power(g, m, p);
+        final long step = power(g, m, p);
         System.out.println("Step value: " + step);
 
-        final int[] bigSteps = new int[m];
+        final long[] bigSteps = new long[m];
         bigSteps[0] = step;
         for (int i = 1; i < m; i++) {
             bigSteps[i] = (bigSteps[i - 1] * step) % p;
@@ -30,14 +30,14 @@ public class SchenksInd {
         System.out.println("Big steps: " + Arrays.toString(bigSteps));
 
         //small steps
-        int currentSmallStep = x; //first step
+        long currentSmallStep = x; //first step
         for (int j = 0; j <= m; j++) {
             System.out.println("Check small step: " + currentSmallStep);
             final int i = arrayIndexOf(bigSteps, currentSmallStep);
             if (i >= 0) {
                 System.out.printf("i = %d%n", i + 1);
                 System.out.printf("j = %d%n", j);
-                final int result = m * (i + 1) - j;
+                final long result = m * (i + 1) - j;
                 System.out.printf("ind(%d, %d, %d) = %d%n", x, g, p, result);
                 return result;
             }
@@ -47,15 +47,15 @@ public class SchenksInd {
         return -1;
     }
 
-    public static int power(final int value, final int power, final int mod) {
-        int result = 1;
+    public static long power(final long value, final long power, final long mod) {
+        long result = 1;
         for (int i = 0; i < power; i++) {
             result = (result * value) % mod;
         }
         return result;
     }
 
-    public static int arrayIndexOf(final int[] array, final int value) {
+    public static int arrayIndexOf(final long[] array, final long value) {
         final int length = array.length;
         for (int i = 0; i < length; i++) {
             if (array[i] == value) return i;
